requirements.txt
apturl==0.5.2
asttokens==2.4.1
autoprop==4.1.0
bayesian-optimization==1.5.1
bcrypt==3.2.0
blinker==1.4
Brlapi==0.8.3
certifi==2020.6.20
chardet==4.0.0
click==8.0.3
cloudpickle==2.2.1
cma==3.4.0
colorama==0.4.6
comm==0.2.2
command-not-found==0.3
contourpy==1.2.1
cryptography==3.4.8
cupshelpers==1.0
cycler==0.12.1
Cython==3.0.10
dbus-python==1.2.18
debugpy==1.8.2
decorator==5.1.1
defer==1.0.6
distro==1.7.0
distro-info==1.1+ubuntu0.2
drake==1.31.0
duplicity==0.8.21
exceptiongroup==1.2.2
executing==2.0.1
Farama-Notifications==0.0.4
fasteners==0.14.1
filelock==3.15.4
fonttools==4.53.1
fsspec==2024.6.1
future==0.18.2
glob2==0.7
glooey==0.3.6
gradescope-utils==0.5.0
gymnasium==0.29.1
httplib2==0.20.2
idna==3.3
imageio==2.34.2
importlib-metadata==4.6.4
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.26.0
ipywidgets==8.1.3
jedi==0.19.1
jeepney==0.7.1
Jinja2==3.1.4
joblib==1.4.2
junit-xml==1.9
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyterlab_widgets==3.0.11
keyring==23.5.0
kiwisolver==1.4.5
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy_loader==0.4
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
manifold3d==2.5.1
manipulation==2024.6.29
markdown-it-py==3.0.0
MarkupSafe==2.0.1
matplotlib==3.9.1
matplotlib-inline==0.1.7
mdurl==0.1.2
meshio==5.3.5
monotonic==1.6
more-itertools==8.10.0
mpld3==0.5.10
mpmath==1.3.0
mutmut==2.5.1
nest-asyncio==1.6.0
netifaces==0.11.0
networkx==3.3
nevergrad==1.0.3
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.82
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.0
olefile==0.46
packaging==24.1
pandas==2.2.2
paramiko==2.9.3
parso==0.8.4
pexpect==4.8.0
pillow==10.4.0
platformdirs==4.2.2
pluggy==1.5.0
pony==0.7.19
prompt_toolkit==3.0.47
protobuf==3.12.4
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycairo==1.20.1
pycups==2.0.1
pydot==3.0.1
pyglet==1.5.29
Pygments==2.18.0
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyMCubes==0.1.6
PyNaCl==1.5.0
pyparsing==3.1.2
pyRFC3339==1.1
pytest==8.3.3
python-apt==2.4.0+ubuntu3
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-fcl==0.7.0.6
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
pyzmq==26.0.3
reportlab==3.6.8
requests==2.25.1
rich==13.7.1
scikit-image==0.24.0
scikit-learn==1.5.1
scipy==1.14.0
SecretStorage==3.3.1
signature_dispatch==1.0.1
six==1.16.0
stable_baselines3==2.3.2
stack-data==0.6.3
sympy==1.13.1
systemd-python==234
threadpoolctl==3.5.0
tifffile==2024.7.24
timeout-decorator==0.5.0
toml==0.10.2
tomli==2.0.1
torch==2.3.1
torchvision==0.18.1
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
trimesh==4.1.8
triton==2.3.1
typeguard==3.0.2
typing_extensions==4.12.2
tzdata==2024.1
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
vecrec==0.3.1
vhacdx==0.0.8.post1
wadllib==1.3.6
wcwidth==0.2.13
widgetsnbextension==4.0.11
xdg==5
xkit==0.0.0
zipp==1.0.0
